+++ buffer_init.c
@@ -1,6 +1,6 @@
 #include "buffer.h"
 
-void buffer_init(buffer *b, int (*op)(int, char *, unsigned long),
+void buffer_init(buffer *b, long (*op)(int, char *, unsigned long),
                  int fd, char *s, unsigned long len)
 {
   b->size = len;
